feat: 상품들의 모든 금액과 배달료를 더하여 상품 금액을 계산한다.

+++ src/main/java/cart/dto/request/OrderRequestDto.java
@@ -3,20 +3,20 @@ package cart.dto.request;
 import java.util.List;
 
 public class OrderRequestDto {
-    private List<Integer> cartItems;
+    private List<Long> cartItems;
     private int totalPrice;
     private int couponId;
 
     public OrderRequestDto() {
     }
 
-    public OrderRequestDto(final List<Integer> cartItems, final int totalPrice, final int couponId) {
+    public OrderRequestDto(final List<Long> cartItems, final int totalPrice, final int couponId) {
         this.cartItems = cartItems;
         this.totalPrice = totalPrice;
         this.couponId = couponId;
     }
 
-    public List<Integer> getCartItems() {
+    public List<Long> getCartItems() {
         return cartItems;
     }
